% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bookdown.R
\name{create_book}
\alias{create_book}
\title{Create an `rstudio4edu` online book with `bookdown`}
\usage{
create_book(dir, title, gh_pages = FALSE, edit = interactive())
}
\arguments{
\item{dir}{Directory for book}

\item{title}{Title of book}

\item{gh_pages}{Configure the book website for publishing using [GitHub
Pages](https://pages.github.com/)}

\item{edit}{Open site index file or welcome post in an editor.}
}
\description{
Create a basic skeleton for an `rstudio4edu` book, using the
`bookdown::gitbook()` output format. Use the `create_book()`
function, or alternately you may use the project wizard in the
RStudio IDE to make a new bookdown site within an R project.
If you want to learn how to make this bookdown site from scratch,
please see:
https://rstudio4edu.github.io/rstudio4edu-book/intro-bookdown.html
}
\note{
The `dir` and `title` parameters are required (they will be prompted for
  interatively if they are not specified).
}
\examples{
\dontrun{
library(rmd4edu)
create_book("mybook", "My Book")
}
}
