% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmdsite.R
\name{create_rmdsite}
\alias{create_rmdsite}
\title{from rmarkdown
and the distill packages}
\usage{
create_rmdsite(dir, title, gh_pages = FALSE, edit = interactive())
}
\description{
from rmarkdown
and the distill packages
}
