% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson.R
\name{lesson_single}
\alias{lesson_single}
\title{Custom HTML template}
\usage{
lesson_single(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Loads additional style and template file
}
