% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distill.R
\name{new_project_create_blog}
\alias{new_project_create_blog}
\title{this is the wrapper for the exported function that is used by
our RStudio project template}
\usage{
new_project_create_blog(dir, ...)
}
\description{
this is the wrapper for the exported function that is used by
our RStudio project template
}
