% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson.R
\name{rstudio4edu_lesson}
\alias{rstudio4edu_lesson}
\title{R Markdown format for rstudio4edu HTML documents}
\usage{
rstudio4edu_lesson(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Loads additional style and template file
}
